+++ src/feature/api/channels/domain/repositories/channel-message.repository.ts
@@ -2,7 +2,7 @@ import { QueryOrder } from '@mikro-orm/core';
 import { EntityManager } from '@mikro-orm/postgresql';
 import { Injectable } from '@nestjs/common';
 import { ChannelMessageEntity } from 'src/feature/api/channels/infrastructure/channel-message.entity';
-import { ChannelMessage } from './channel-message';
+import { ChannelMessage } from '../channel-message';
 
 @Injectable()
 export class ChannelMessageRepository {
